requirements.txt
Django==5.0.6
mysqlclient==2.2.4
python-decouple==3.8
python-dotenv==1.0.1
django-crispy-forms==1.14.0
